address review comments

+++ frame/authority-discovery/src/lib.rs
@@ -131,7 +131,7 @@ impl<T: Config> OneSessionHandler<T::AccountId> for Pallet<T> {
 		}
 	}
 
-	fn on_disabled(_i: usize) {
+	fn on_disabled(_i: u32) {
 		// ignore
 	}
 }
@@ -233,7 +233,7 @@ mod tests {
 		) {
 		}
 
-		fn on_disabled(_validator_index: usize) {}
+		fn on_disabled(_validator_index: u32) {}
 
 		fn on_genesis_session<Ks: OpaqueKeys>(_validators: &[(AuthorityId, Ks)]) {}
 	}
